Create desktop.py
